memset_base(): avoid warnings in stricter C++ compiler
If we try to compile memset.c with a C++ compiler, we get warnings on
various implied pointer conversions. Let's make them explict, and this
way the code can compile as either C or C++ code.

I want to be able to compile it as C++ code because in the next patch
I want to use it in fastlz/lzloader.cc.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>