tracepoints: fix build error in debug mode on recent gcc
This patch fixes a link error in debug mode with gcc 8.3.1 (e.g, Fedora 29).
The error lists several cases of tracepoint code which refers to a
"discarded section".

Good practice suggests that tracepoints be defined in source files, not in
header files. The issue is that when a (static) tracepoint is defined in a
header file, multiple copies of it are created, one for each of the source
files which include this header file. This is in fact legal and we support
it, but the separate copies waste memory and discouraged.

Nevertheless, in some rare cases we did decide to have a tracepoint in the
header file - in particular when inline code in that header file needs to
use a tracepoint. But when we do this, the function that uses the
tracepoint *must* be inlined. If the compiler decides not to inline it,
several copies of the function will be compiled for different source
files, and these copies are *not* identical (each refers to a different
copy of the tracepoint variable!) but the linker will assume they are,
and discard all of the copies except one.

So this problem happens when the compiler decides to *not* inline an
inline function. Apparently in gcc 8.3.1's debug mode, not inline is
more common than it used to be, so we started seeing this problem.

The easy workaround is to use __attribute__((always_inline)) to force
inlining of the problematic functions, in any case. Doing this fixes
the debug build.

This workaround still leaves us in violation of the ODR ("one definition
rule"), i.e., multiple translation units define the same method in
different ways (since each writes to a different tracepoint copy).
This violation is what caused this bug in the first place. So a cleaner
solution would be to eliminate the ODR violation. In other words,
do not define tracepoints in header files at all, and rather define
them in a source file and only "extern" them in the header file.
But let's leave that for a future patch.

Fixes #1029.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>