Make RAMFS not to free file data when file is still opened
Even though it is valid to delete a file its data (i-node)
should not be deleted until all file descriptors are closed.

This patch fixes the bug in file deletion logic in RAMFS
to make sure that file node does not get deleted
until all file descriptors are closed.

Fixes #1035

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>