tracepoints: fix for compiling on gcc 9
In gcc 9.1.1, strictly following the C++ standard, one is no longer
allowed to specify attributes, e.g., [[gnu::cold]], on function
definitions - only on function declaration. As an annoying side-effect,
we get warnings if we specify them on a lambda, which is both a declaration
and a definition.

The easiest workaround is to use the old-style __attribute__((cold))
instead of the more modern syntax [[gnu::cold]]. The old-style syntax
isn't bound by the C++ standard, and its behavior hasn't changed in
gcc 9.1.1.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190519122628.6226-1-nyh@scylladb.com>