Change vmlinux_entry64 to switch to protected mode and jump to start32
This patch changes vmlinux_entry64 to switch to protected (32-bit) mode
and jump to the start32 in order to setup transition to long mode
in a single consistent way.

In essence this patch removes unnecesary duplication between
vmlinux-boot.S and boot.S. But more importantly it corrects
the vmlinux boot code (used with firecracker) by making sure
that all data segment registers (ES, DS, etc) are setup correctly to point
to OSv GDT table (see start32) which is critical when kernel
is not mapped 1:1.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20190615034622.23154-1-jwkozaczuk@gmail.com>