Enhance scripts/test.py to allow running unit tests on firecracker
This patch enhances scripts/test.py to allow running unit tests on firecracker.
It does so by adding new parameter: -p HYPERVISOR.

Please note that the scripts/build does not have an ability to pass
this new parameter to scripts/test.py when invoking it. So in order to execute tests
on firecracker one has to build test image first and then manually invoke test.py like so:

./scripts/build image=tests
./scripts/test.py -p firecracker

Finally please not that on first test run, the firecracker.py
script will ask you to enter root password to allow setting up a TAP device
needed to enable networking with firecracker.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>