options: implement light and simple alternative to boost program options
This patch provides simple and light alternative to boost program options
library. It adds handful of utility functions that are aimed
to help parse command line options and intended to be used
by kernel loader and the utility apps like cpiod, httpserver
and cloud init.

For more details on what particular functions do
look at the comments in options.hh and options.cc.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20191003042437.27978-2-jwkozaczuk@gmail.com>