Support local-exec TLS access
This patch enhances OSv dynamic loader to support
pies and position dependant executables that use TLS
(Thread Local Storage) in local-exec mode.

It does so by reserving an extra slot in kernel static
TLS block at its end and designating it as user static TLS
for the executable ELF. Any dependant ELF objects are still
placed in the area before the kernel TLS. For the specifics please
read comments added to arch-elf.cc and arch-switch.hh.

Please note that this solution limits the size of the application
TLS block to 64 bytes plus extra gap due to 64-bytes alignment
of the kernel TLS. This should be sufficient for most applications
which either use tiny TLS (Golang uses 8-bytes long) if at all.
Rust ELFs tend to rely on quite large TLS in which case the limit
in loader.ld needs to be increased accordingly and loader.elf
relinked.

Fixes #352

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>