virtio: adjust virtual queues initialization to adhere to the suggestions in the spec
This patch changes slighly how the virtual queues get detected
and activated.

Before this patch the all the queues would be first detected in first loop
and the activated in the next one. This not how the virtio spec suggests it,
especially for virtio-mmio devices:

"4.2.3.2 Virtqueue Configuration
The driver will typically initialize the virtual queue in the following way:

1. Select the queue writing its index (first queue is 0) to QueueSel.
2. Check if the queue is not already in use: read QueueReady, and expect a returned value of zero (0x0).
3. Read maximum queue size (number of elements) from QueueNumMax. If the returned value is zero (0x0) the queue is not available.
4. Allocate and zero the queue pages, making sure the memory is physically contiguous. It is recommended to align the Used Ring to an optimal boundary (usually the page size).
5. Notify the device about the queue size by writing the size to QueueNum.
6. Write physical addresses of the queue’s Descriptor Table, Available Ring and Used Ring to (respectively) the QueueDescLow/QueueDescHigh, QueueAvailLow/QueueAvailHigh and QueueUsedLow/QueueUsedHigh register pairs.
7. Write 0x1 to QueueReady."

This patch changes the logic so that the queues are detected and activated in single loop. This way
OSv can properly boot on qemu microvm machine.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>