mempool: use map_anon() for large allocations or when memory is fragmented
This patch mostly addresses the issue #854. In essence it changes malloc_large()
to use mmu::map_anon() if requested memory is greater than size of "huge page"
and it does not have be contiguous in physical memory. Also it makes it fallback
to map_anon() based allocation for requests > 4K and < 2MB that cannot be satisfied
when memory is heavily fragmented.

It is supposed to help OSv memory allocation behave better when memory
in free_page_ranges is heavily fragmented and large allocations (>=2MB)
cannot be satisfied with single contiguous page range. When working on
this patch it has also become clear, that applications that heavily use realloc()
which allocates new chunks of memory and frees old ones, might very much encounter
heavy fragmentation of memory. Hopefully this patch is going to help
in these scenarios.

It just happens that object_size() and large_object_size() needed to be
adjusted to work with new malloc_large() implementation. It was discovered
that therefore other related issues are fixed by this patch as well.

Lastly this patch adds new test verifying realloc() and malloc_usable_size().

Fixes #784
Fixes #854
Fixes #1077

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>