socket: warn about missing ancillary information handling
Nadav Har`El writes:
"I think we never properly implemented the "ancillary data" (control message) feature of sockets which can be used to, among other things, send an open file descriptor between two processes communicating over a UNIX-domain socket. I don't remember why we never implemented this - I think the problem was (considering the file this is in, compat/linux/linux_socket.cc) that the BSD-based code assumes this ancillary data has one structure, but the Linux API assumes a different structure. Unfortunately this will be hard to fix properly without looking at the code and at the socket manual pages, and trying to remember what the heck this code is supposed to be doing. I also see there's a lot of #if 0 code in this area :-(

Perhaps we can take a patch which doesn't outright remove these asserts, but just replaces them by a WARN_ONCE or something, so we'll allow these control messages but warn that they may not be handled properly."

This patch makes it possible to run som XWindows apps like xclock, xeyes, etc.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>