[PATCH 3/4] arm on firecracker: clean cmdline from early console artifacts
In order to enable early console on firecracker, we have
to prepend the command line with 'console=...' name,value pair.
Firecracker then appends extra name,value pair containing
memory address of the console base register in form of 'earlycon='

We actually do not need to parse this information from command line,
instead we get it from DTB which one of the previous patches
addresses. So all in all, the only purpose of this patch
is to remove this irrelevant information if present so that
command line contains normal OSv boot parameters.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>