scripts/build: refine build process for aarch64
Since we do not create usr.img artifact for aarch64, we should
not try to set command line in `scripts/build` if usr.img does not exist.

Also let us export the STRIP environment variable to point to aarch64
version of the strip utility so that scripts/module.py and related
python scripts use the version matching the achitecture.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>