aarch64: support building unit tests image
This patch mostly modifies module tests makefile to support
building unit tests image for the aarch64 architecture. Please note
that handful of tests are not get built for aarch64 for all kinds of
reasons (please see the comments in the tests makefile).

At this moment it is possible to build a rofs test image only (the zfs
file system is not functional at this point) like so:

`./scripts/build image=tests fs=rofs -j4 arch=aarch64 --create-disk`

Until we make ./scripts/test.py arch-aware, one needs to execute a test
manually like so:

`./scripts/run.py -c 1 --arch aarch64 -V -e '/tests/tst-chdir.so'`.

It looks like many tests are passing, but good chunk are not, mostly
because of lack of floating point support and other missing
features in aarch64 port.

Finally, please note that the aarch64 test image contains the java tests
with x64 jvm binaries which obviously will not work on aarch64. For now
we should somehow disable adding to the aarch64 test image.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>