tests: fix compilation of tst-wctype.cc on older versions of gcc
For whatever reason tst-wctype.cc compiles fine on gcc 10
even though it uses std::cout but does not include <iostream>
header. Older versions of gcc (or there is other reason)
are not as lenient.

So this patch simple adds missing include to make this file
compile on older versions of Fedora.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>