loader: add support for booting from virtio-fs
This makes the necessary and pretty straight-forward additions to the
laoder to support using virtio-fs as the root filesystem. It also makes
minimal changes to scripts/build to add support there as well. Note
that, to obtain a directory with contents specified by the manifest
files, usable as the virtio-fs host directory, one can use the existing
'export' and 'export_dir' (previously undocumented) options to
scripts/build.

Ref https://github.com/cloudius-systems/osv/issues/1062.

Signed-off-by: Fotis Xenakis <foxen@windowslive.com>
Message-Id: <AM0PR03MB62921196B1CC7E09D5EFA29DA62C0@AM0PR03MB6292.eurprd03.prod.outlook.com>