libc: drop 5 more headers with symlinks to musl copies
Those 5 headers are mostly identical to their musl copies
except they were modified to comment out *64 macros that serve
as aliasing mechanism on musl side but conflicts in all
places where we use LFS64 macro that uses weak_alias to provide
symbol copies.

So in order to avoid this collisions and replace the headers
with symlinks we drop LFS64 in favor of aliases.ld mechanism
which achieves the same goal.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>