tests: enhance test.py to support running unit tests on aarch64
This adds new parameter to test.py - '--arch' - which can
be used to pass the same parameter to run.py in order to run
unit tests image on specified architecture with qemu.
This is mostly aimed to support running unit tests
on aarch64.

This patch also adds new list of tests that are to be disabled on aarch64
until various aarch64 related shortcomings in OSv are addressed.

Here is how you can build ROFS test image and run it using test.py.

./scripts/build arch=aarch64 -j4 fs=rofs --create-disk image=tests
./scripts/test.py --arch=aarch64 --run_options="-c 1"

Please note that we pass "-c 1" to force running all tests
with single vCPU as some tests simply fail or hang with more vCPUs.
At this moment 130 unit tests pass on aarch64 (most of the time).

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>