syscall_to_function.h: simplify, and avoid C++20 extension
The macros in syscall_to_function.h used "__VA_OPT__", which is a C++20
extension that older compilers (such as on Ubuntu 18.04) don't understand.

Luckily, we don't even need to use this extension. The macros in this file
used an elaborate way to change __OSV_TO_FUNCTION_SYS_fcntl to just "fcntl"
with all the parameters intact. But the easiest way to achieve that goal
is just to #define the name - without even telling the C preprocessor that
this is a function! It will just change "__OSV_TO_FUNCTION_SYS_fcntl" to
"fcntl" and not worry at all about anything which may or may not appear
afterwards in parentheses. It's easier, shorter, and not C++20-specific.

Fixes #1099

Signed-off-by: Nadav Har'El <nyh@scylladb.com>