aarch64: fix memory mapping initialization when booting with #vCPUs >= 2
As the issue #1092 explains, OSv hangs most of the time when booting
with 2 or more vCPUs. More specifically it only happens in non-emulated
mode on QEMU or Firecracker with KVM acceleration on real hardware host
like Raspberry PI 4. It is also important to mention that very
occasionally OSv successfully boots with more than one vCPU in
such setup.

Based on the results of the remote debugging session, it seems
that OSv hangs on primary vCPU in the `smp_launch()` routine
while waiting for all secondary vCPUs to come up. More specifically
the stack backtrace on the secondary cpu seems to indicate
an exception raised in `init_mmu` function in `boot.S`
at line 161 (https://github.com/cloudius-systems/osv/blob/0af6dd744817c847878e498a0ae34219fc5bf613/arch/aarch64/boot.S#L161)
which is executed right after `init_secondary_pt`. This would
mean that secondary vCPU crashed. The secondary cpu registers
`ttbr0_el1` and `ttbr1_el1` which should hold addresses of
the translation table roots are set to 0 which seems to be incorrect
and most likely caused the crash in `init_mmu`. Both
`ttbr0_el1` and `ttbr1_el1` registers are prepared to be set in
`init_secondary_pt` which copies values from `smpboot_ttbr0`
and `smpboot_ttbr1` global variables.

Both `smpboot_ttbr0` and `smpboot_ttbr1` variables are set in
in the loop in `smp_launch()` for every secondary vCPU (which seems
wrong as it should be done only once). It seems that when
a secondary vCPU boots and calls `start_secondary_cpu`, it does
not see the correct values of these global variables in memory possibly
due to some memory coherency side effects.

So this patch makes both `smpboot_ttbr0` and `smpboot_ttbr1`
set much earlier exactly just when the runtime translation tables
switch is executed on the primary cpu in the
`switch_to_runtime_page_tables()` function. This change by itself
makes OSv boot most of the time with 2 or vCPUs but it still hangs
sometimes. This most likely can be explained by the fact that
`start_secondary_cpu` directly uses runtime tranlation tables
which resides in memory not visible by default. To counter
it, this patch also changes `start_secondary_cpu` in boot.S
to first enable the boot translation tables and only then switch
to the runtime tables (see added calls to `init_boot_pt` and
`init_mmu`).

Based on the empirical tests (more than 100 executions) OSv can now boot
with 2 or vCPUs with KVM acceleration on.

Fixes #1092

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20201009210208.461186-1-jwkozaczuk@gmail.com>