aarch64: store ESR (Exception Syndrom Register) in the exception frame
As I was chasing another bug, I have discovered that the synchronous
exception handler - entry_sync - was not saving the state of ESR_EL1
as part of the exception frame even though there was already a field for it.

The handler entry_sync was acting correctly on the value of ESR_EL1, but because it was
not being saved as part of the frame the downstream logic in arch/aarch64/mmu.cc
and core/mmu.cc relying on the value returned by get_error()
method, would receive bogus value and occasionally crash with errors
like 'page fault outside application ...' due to wrong permissions.

This patch fixes it but adding single line of assembly in entry.S
to save ESR_EL1 under esr field of the frame.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20201122052644.168472-2-jwkozaczuk@gmail.com>