build: enforce Partial RELRO for internal apps
On most Linux distribution like Ubuntu or Fedora gcc links shared
libraries and executables with Partial RELRO. On others like Nix OS
the Full RELRO mode is used. In essence in Partial RELRO mode certain
parts of ELF segments like GOT are left writable after processing
headers thus allowing so called lazy resolving of symbols later. On other
hand, in Full RELRO mode all symbols are resolved eagerly and the GOT
part is disabled for writing after. For more details please read
this article -
https://blog.osiris.cyber.nyu.edu/exploitation%20mitigation%20techniques/exploitation%20techniques/2011/06/02/relro-relocation-read-only/.

Both modes are supported by OSv dynamic linker and in Full RELRO mode
OSv actually ignores symbols not found when resolving them eagerly.

Some internal applizations like mkfs.so and zfs.so depend on
common library libzfs.so. Weirdly, libzfs.so also uses some symbols
from both mkfs.so and zfs.so. Because of this, when all aforementioned
apps are linked in Full RELRO it creates a problem when they are
executed sequentially in the same OSv instance which is the case when
building ZFS image.

More specifically when mkfs.so with Full RELRO is executed first,
all symbols in mkfs.so and libzfs.so that can be found at this point
are resolved eagerly and GOT for both is made read-only after.
As an example, one of the symbols referenced by libzfs.so is
bsd_getmntany. The bsd_getmntany is part of zfs.so which has not been
loaded yet at this point so its reference in libzfs.so is left unresolved.
When finally zfs.so is loaded, OSv resolves
all symbols in zfs.so eagerly and tries to run it. Unfortunately when
during this process bsd_getmntany in called by some code in libzfs.so it
triggers resolve_pltgot() which finds the bsd_getmntany symbol now but
fails with page fault as GOT in libzfs.so is marked as read-only.

It is not clear if this situation is a real-world scenario and
would apply to any real Linux app. If that is the case it would
mean that OSv dynamic linker has the limitation to handle such
scenarios.

So for now we are explictly telling gcc to use Partial RELRO
mode for mkf.so, zfs.so and other internal apps and libraries
regardless of what defaults are on given Linux distribution.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>