Add test for fread() hang when reading from a directory
In commit c622cf7b8fb7b88188032165e0ee035e9924b71b we fixed a bug where
fread() on a directory hangs, instead of resulting in an error as expected.
However, we didn't have a test for this case - and this patch adds one.

The new test hangs before the aforementioned commit, and passes without
it. The test also passes on Linux.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20201229135830.3414999-1-nyh@scylladb.com>