bootchart: do not print empty entries
On aarch64 the first 2 bootchart entries are never recorded
as they are specific to x86_64 and have to do with measuring
16- and 32-bit phases of the boot process. Unfortunately,
it causes them to be printed with '(null)' when --bootchart
option added to command line.

This patch simply changes boot_time_chart::print_one_time()
to ignore empty entries.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>