aarch64: make boot messages quieter and consistent with x86_64
Most likely in the early stages of the aarch64 port development
it was valuable to print debug messages. Unfortunately, this creates
a lot of noise in the console output especially when every page fault
gets extensively loggged.

So this patch simply disables each relevant debug statement by
using '#if CONF_logger_debug' macro so that it would still get printed
in debug build.

This also makes the console output more consistent with what we print
on x86_64 with '--verbose' flag on or off.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>