aarch64: implement processor_to_nano in armclock
This patch provides missing implementation of processor_to_nano()
in the arm_clock driver.

We simply re-use the formula in the arm_clock::uptime() method
and instead of reading the counter, we simply apply the passed
in one and convert to nanoseconds. This is enough to make bootchart
work properly on aarch64.

Partially addresses #1091

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>