aarch64: Support downloading packages on Ubuntu 18.04 x64
The codename returned by linux_distribution() on Ubuntu 18.04 was
'Bionic Beaver', but we want 'bionic'. Switch to
distro.lsb_release_attr('codename') to grab the correct form of the codename.

Also, fixup the URL generated for downloading boost packages to include the
patchlevel of the boost version, but only if the patchlevel is non-zero.

I tested this on Ubuntu 18.04, Ubuntu 20.04, and Fedora 31.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>
Message-Id: <20210116155126.6760-1-stewart.hildebrand@dornerworks.com>