modules: add common makefile to allow cross-compiling
Currently there are many variables and rules repeated across makefiles
of individual modules under the ./modules subdirectory. To make things
worse, most of the modules except tests do NOT support cross-compiling aarch64.

To address the above, this patch adds common makefile -
./modules/common.gmk - intended to be included by the makefiles of
individual modules to re-use common rules and variables.

More specificaly, common.gmk, similarly to the main makefile, discovers
the host architecture and uses cross-compiler prefix if applicable to
identify directories of standard and boost C++ headers as well as
location of shared libraries like libstdc++.so and libgcc_s.so.1.
Similarly to the main makefile, this logic should work on both native
x86_64 and aarch64 host as well as when cross-compiling aarch64 on x64
on Ubuntu and Fedora.

Eventually, the common compiler flags including the include headers
are set in the COMMON variable.

Fixes #1105

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>