build: set arch in arg-parsing loop
Add a case in the initial argument parsing loop to both set the variable and
append it to the args array.

In the follow up patch, there will be a need to know if we are doing a cross
build. In this case, passing arch=aarch64 is required, thus the arch= variable
will be set in the initial arg parsing loop.

If arch= isn't given on the build command line, it will be unset initially.
That's okay, since we only need the arch variable early if it's explicitly given
on the build command line. The arch variable will still be set later in the
script regardless.

While we're here, use a cleaner syntax for appending to an array and fixup an
outdated comment.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>

Message-Id: <20210208184545.97466-2-stewart.hildebrand@dornerworks.com>