Add script to download aarch64 cross compiler toolchain
This is a source-able script that downloads the aarch64 toolchain from
developer.arm.com, extracts it, and, when being sourced, adds the path to PATH.
If ccache is installed, it will also add the ccache directory to the front of
PATH.

Add a check in the OSv build script to source the toolchain if arch=aarch64 and
the toolchain has been downloaded to the build directory.

Add "centos" as an option in download_aarch64_packages.py to download the
toolchain. I did not change the logic for Ubuntu here due to the possibility of
ending up with unvetted combinations of this GCC version with various Ubuntu
system libraries that could be built with different versions of GCC. This
toolchain can still be used on Ubuntu by sourcing it directly, but in this case
I recommend building Boost and other dependent libraries from source to ensure
they are built with the same toolchain that is being used to build OSv and your
apps/modules.

This toolchain should work on any distribution that ARM supports, which is
Ubuntu 16.04 LTS or later, or RHEL 7 or later. I've tested this on CentOS 7 and
Ubuntu 18.04.

My primary motivation for this is to support the aarch64 build on CentOS 7.
While the EPEL repository for CentOS 7 does have an aarch64 cross compiler
toolchain available, it is quite an old version (4.8.5). On Ubuntu, the aarch64
cross compiler toolchain conflicts with the gcc-multilib package. Using the ARM
toolchain allows us to avoid these particular issues.

This toolchain does not provide Boost, so Boost will need to be built. Boost can
be built with the script download_and_build_boost.sh in the scripts/ directory,
or it can be built manually. Here are instructions I used for building Boost
from source. Ensure that the current working directory is the root of the OSv
repository, then run the following commands:

  $ source scripts/download_aarch64_toolchain.sh
  $ mkdir -p build/downloaded_packages/aarch64/boost
  $ pushd build/downloaded_packages/aarch64/boost
  $ wget https://dl.bintray.com/boostorg/release/1.70.0/source/boost_1_70_0.tar.gz
  $ tar -xf boost_1_70_0.tar.gz
  $ cd boost_1_70_0/
  $ BOOST_DIR=$(pwd)
  $ ./bootstrap.sh
  $ echo "using gcc : arm : aarch64-none-linux-gnu-g++ ;" > user-config.jam
  $ ./b2 --user-config=${BOOST_DIR}/user-config.jam toolset=gcc-arm architecture=arm address-model=64 -j$(nproc)

After building Boost, we need to set up some symlinks so that OSv can find the
Boost headers and libraries. Ensure your current working directory is the Boost
source directory BOOST_DIR (it should be already), then run the following
commands:

  $ ln -s boost_1_70_0/stage ../install
  $ mkdir -p stage/usr/include
  $ ln -s ../../../boost stage/usr/include/boost
  $ popd

After building Boost and setting up the symlinks, we should be ready to build
OSv with the aarch64 toolchain:

  $ scripts/build arch=aarch64 fs=ramfs --create-disk image=native-example -j$(nproc)

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>

Message-Id: <20210208184545.97466-3-stewart.hildebrand@dornerworks.com>