Introduce Cadence UART driver
The irq bits need to be cleared after reading the fifo register, so the fifo
register needs to be read from interrupt context. Therefore, I introduced
the variables _input_ready and _uart_fifo to store the RX character during the
interrupt handler to be retrieved by the service thread later.

The driver will be initialized from device tree in a follow-up patch.

I modeled the structure of the driver roughly after the OSv pl011 driver.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>
Message-Id: <20210311171431.31417-2-stewart.hildebrand@dornerworks.com>