tests: improve makefile by using more portable case statement
The tests makefile uses the shell if statement to detect CROSS_PREFIX
in order to use appropriate method of identifying boost dependencies.
Unfortunately this only works if the shell used by make is bash.
To fix this we replace if by case statement which also works in default
shell used by make.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>