aarch64: make kernel_vm_base equal to kernel_base
Until we implement #1087 to dynamically map kernel ELF
located anywhere in physical memory, we always need to make
sure that that the kernel_base is eqaul to kernel_vm_base in
the makefile. So let us explicitly state this in rule in the
makefile.

The extra benefit is that it should be now a bit easier to build
kernel for hypervisors that load kernel ELF in a place different
than default QEMU one - 0x40080000.

For example to build loader.img for firecracker it should be as easy
as:

pushd build/release.aarch64/ && rm -f loader.* arch/aarch64/arch-dtb.* core/elf.* core/mmu.* && popd \
&& ./scripts/build -j$(nproc) image=native-example fs=rofs mode=release --create-disk kernel_base=0x80080000

./scripts/firecracker.py -k build/release/loader.img -i build/release/disk.raw -c 2

Ideally at some point in future we could enhance the makefile
to make it generate some sort of header file with variables
like kernel_base and have it be included by source files (core/mmu.cc,
etc) that depend on it. That way simply passing kernel_base parameter
to the build command would automatically rebuilt all necessary sources
without having to explicitly delete relevant object files which
we have to do now.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>