setup.py: add readelf/grep
These are pre-requisites needed for the next patch. Specifically, the next patch
needs the following shell utilities (packages):

  readelf (binutils)
  grep (grep)

The next patch also needs awk, but that utility is part of POSIX and is
installed by default on all supported distributions, so no need to install it
here.

Note that these are not actually new dependencies, this is just making them
explicit. readelf, grep, and awk are already used elsewhere in the Makefile.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>
Message-Id: <20210317182712.106035-2-stewart.hildebrand@dornerworks.com>