aarch64: populate image size in header
Depending on where the bootloader places loader.img and the DTB, there can be a
conflict resulting in the the zero_bss loop in boot.S overwriting the DTB.

I observed this with certain ramfs sizes when building OSv for LynxSecure
hypervisor. For example, the bootloader happened to load the DTB at address
0x01080000, while the end of the .bss and tls/tcb sections (up to .edata) was
0x0108c000, resulting in the zero_bss subroutine zeroing out the dtb. I could
also reproduce this with qemu, by adding a large ramfs resulting in a loader.img
with the approximate size of 127.27MB. In this example, qemu placed the DTB at
address 0x48000000, the kernel at address 0x40080000, and the end of .bss was
0x4805c000.

The bootloader doesn't have awareness of the elf sections, so it can't account
for the .bss section and other sections that have no need to take room in the
elf file due to being the NOBITS type.

Fix this by populating the arm64 image size field in the OSv kernel header. The
size includes the .bss section, thus the bootloader should reserve the area and
avoid overlapping with the DTB.

Add a case in imgedit.py as a helper function.

https://www.kernel.org/doc/Documentation/arm64/booting.txt

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>
Message-Id: <20210317182712.106035-3-stewart.hildebrand@dornerworks.com>