aarch64: make synch_thread/synch_port work with weak memory model
This patch changes bsd/porting/synch.cc to change _awake variable to an atomic
for similar reasons as the t variable in the waiter. In this case, the synch_port::wakeup*()
methods use "void thread::wake_with(Action action)" and we need to make sure the _awake variable
is visible between CPUs.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>