libc: add strerrordesc_np
Per the specification, strerrordesc_np() is supposed to
return a pointer to a string that describes the error code
passed in the argument errnum just like strerror() does.
However unlike strerror(), strerrordesc_np() is NOT supposed
to apply any translation per current locale.

But the OSv implementation of strerror() does not apply any
translation anyway (please see the LCTRANS macro in locale_impl.h
that returns the msg "as is" without any translation). So given
that we simply make strerrordesc_np an alias of strerror.

The strerrordesc_np() is needed by newer version of nginx.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20210525050253.1141211-1-jwkozaczuk@gmail.com>