futex: implement FUTEX_WAIT_BITSET
New version of C++ standard library that is part of GCC 11 toolchain
started using futex() syscall with the FUTEX_WAIT_BITSET operation.

It turns out we are pretty lucky (so far at least) because the bitmask value
passed in the val3 parameter is equal to FUTEX_BITSET_MATCH_ANY. According to the spec,
it means that FUTEX_WAIT_BITSET with val3 equal to FUTEX_BITSET_MATCH_ANY is
equivalent to FUTEX_WAIT except for the different handling of the timeout argument.

This patch implements minimal functionality of FUTEX_WAIT_BITSET that
handles FUTEX_BITSET_MATCH_ANY and is enough to make failing
tst-async.cc and tst-promise.cc pass.

Fixes #1147

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20210621060640.8041-1-jwkozaczuk@gmail.com>