aarch64: on Fedora 34 download the aarch64 artifacts from the version 33
For some reason, the gcc-c++-aarch64-linux-gnu package Fedora 34 ships with,
installs the version 10.2.1 of gcc aarch64 cross-compiler which is different
from the regular x86_64 gcc compiler - 11.1.1.

To avoid compilation errors caused by the mismatch between compiler and aarch64
C/C++ headers Fedora 34 provides, this patch modifies the download_aarch64_packages.py
script to download all relevant artifacts from Fedora 33 instead if the version of
cross-compiler is 10.2.1. The version of cross-compiler on Fedora 33 and
34 is exactly the same, therefore the C/C++ headers downloaded from
version 33 are compatible and can be used with the cross-compiler on
Fedora 34.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>