replace AARCH64_PORT_STUB with __aarch64__ and __x86_64__
As Nadav Har'El noted, the AARCH64_PORT_STUB has been abused
and the original intention of this macro, as its name suggests,
was to mark code that was done as a quick hack or stub for aarch64.

To that end this patch replaces "#ifdef AARCH64_PORT_STUB" with
"#ifdef __aarch64__" and "#ifndef AARCH64_PORT_STUB" with
"#ifdef __x86_64__".

Signed-off-by: wkozaczuk <jwkozaczuk@gmail.com>