in_pcb: extract union in_dependaddr
This patch is a small subset of the original FreeBSD one -
https://reviews.freebsd.org/rS334719 - to add support of
the new SO_REUSEPORT_LB socket option.

The changes below do not affect any functionality, but instead
simply extract new union in_dependaddr that will be used
in the load balancer group struct defined as part of the
forth coming patch.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20210822042314.167929-1-jwkozaczuk@gmail.com>