expose inet_aton and inet_ntoa when kernel built with most symbols hidden
When we compile OSv with most symbols hidden, most of the symbols
defined under the bsd/ subtree get hidden. This includes the inet_aton()
and inet_ntoa() functions that are internally used in the FreeBSD-based
networking stack implementationb. This normally would be fine and desired
as OSv also links with the user-facing version of one of these functions
- inet_aton() - see musl/src/network/inet_aton.c. There is also musl version
of inet_ntoa() under musl/src/network/inet_ntoa.c which currently OSv
does not build with.

It turns out however, that musl/src/network/inet_aton.c implements
'__inet_aton' and exposes it as `inet_aton` using weak_alias macro. This somehow
gets silently ignored when linked and we seem to be getting the BSD version of inet_aton
exported. The FreeBSD version seems to be quite different from the musl one
so maybe we should hide it by renaming the BSD version of inet_aton and
thus properly expose the musl version.

The musl and FreeBSD version of inet_ntoa seems to be almost identical
so this is not a big issue. However it would be ideal if we also
exported the musl version of this function.

For now until we decide what is the best solution, this patch exposes
the bsd versions of inet_aton and inet_ntoa so that when kernel is built
with most symbols hidden, these two ones are visible.

Refs #1178

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>