linker: enforce some sections kept when garbage collection on
This patch changes the x64 linker script to keep some sections
when garbage collection (--gc-sections) is on when linking the kernel.

This mostly involves adding KEEP directives to specify number of extra
entry points - start32_address, start32_from_vmlinuz_address - so that
linker treats them as extra code roots and does not eliminate relevant code
from the resulting elf file.

Similarly we also add KEEP to .memcpy_decode, .fixup, .eh_frame*,
.tracepoint_patch_sites* and .init* sections to make them retained
when linking with garbage collection on.

This changes to not have any effect on the linker result when garbage
collection is off - the default.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>