java: include the conf/ subdirectory of openjdk 9 and above
This patch modifies the makefile of openjdk9_1x-from-host module
to include the conf/ subdirectory which among many things contains
key configuration files like java.security that drives random generation
logic and affects creating temporary files. If those files are missing
creating temporary files is very slow as described in the issue #1165.

Besides addressing the isssue #1165 this patch adds new Java test
to verify that the temporary files can be created correctly on JDK 9 and
above. It also makes both - 8 and 9+ - JDK modules include libnss3.so
which is necessary with the version of JDK on Fedora.

Fixes #1165

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>