exported symbols: enhancements
This patch simplifies how lists of exported symbols are stored
under exported_symbols/ directory in order to ease the maintainance
of those as new symbols will be added in future.

In essence we store symbols that are common to both architectures -
x86_64 and aarch64 - in the files exported_symbols/osv_*.symbols. The
symbols that are unique to given architecture are stored under
exported_symbols/x64/ or exported_symbols/aarch64. This allows us
to minimize number of files we need to add new exported symbol to -
osv_ld-musl.so.1.symbols and one of the other osv_*.symbols files.

It is also worth to note that there is no practical significance
which of the files given symbol is located in as all of the symbol files
are merged into build/release/version_script fed to a linker. So as
long as the symbol is located in at least one of symbol files, it will
be exported. However, which file given symbol is part of, will help with
glibc/musl API documentation down the road.

Two accommodate these changes we also modify two script files:
generate_version_script.sh and extract_symbols.sh.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>