libc: implement _dl_find_object()
Ubuntu 2022.04 comes with a new version of GCC 11.2.0 that
somehow includes instance of libgcc_s.so.1 destined for GCC_12.0.0
at least based on what readelf shows. The implication of this is
that during exception handling and stack unwinding, this version
of libgcc_so.so.1 uses _dl_find_object() function what was
very recently added to glibc. For more details please read
following:
- https://www.mail-archive.com/gcc-patches@gcc.gnu.org/msg275982.html
- https://www.mail-archive.com/gcc-patches@gcc.gnu.org/msg273082.html
- https://github.com/gcc-mirror/gcc/commit/790854ea7670f11c14d431c102a49181d2915965
- http://www.gnu.org/software/libc/manual/html_node/Dynamic-Linker-Introspection.html

So this patch adds basic (a little incomplete) implementation of _dl_find_object()
that can satisfy the need of new libgcc_s.so.1 - field dlfo_eh_frame of the struct
dl_find_object. Please note that for now we do not populate the dlfo_link_map field
as it is not clear what exactly goes it there and how it is used. We may need to
revisit this later.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>