netlink: set LINUX_RTM_NEWADDR and LINUX_RTM_NEWNEIGH on responses
This patch fixes a minor bug in handling RTM_GETADDR and RTM_GETNEIGH
requests. It tweaks the relevant code to set the RTM_NEWADDR and RTM_NEWNEIGH
type for the responses respectively.

This is important as for example Golang runtime tests the nlmsg_type of
the netlink response and breaks if it is wrong.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>