netlink: IFA_ADDRESS needs to go first
Golang uses the netlink interface RTM_GETADDR to query the network interfaces and IPs.
It assumes that the 1st attribute in the RTM_NEWADDR response is IFA_ADDRESS. This
patch changes the order in which RTM_NEWADDR attributes are sent to make
sure the IFA_ADDRESS goes first and IFA_LABEL last.

This does not seem to be documented anywhere but Linux sends RTM_NEWADDR responses
with the IFA_ADDRESS attribute first so we follow suit.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>