bsd: disable false-positive warning
Gcc 12.1.1 started complaining on bsd/sys/kern/subr_taskqueue.c's
taskqueue_run_locked() that a pointer to an on-stack object (tb) is saved
on a heap variable (queue).

However, I believe this is a false positive: This code indeed adds a
local object to a queue, but before returning from the function, it
removes this local object from the queue, so the pointer to the stack
cannot remain after this function returns. I don't know why gcc thinks
it can prove that it can remain.

So in this patch I simply disable "dangling-pointer" checking on this
source file.

This is one of the fixes needed to build on Fedora 36 (refs #1198).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>