libc: fix build of qsort() and qsort_r()
Musl's implementation of qsort() and qsort_r() (in newer musl, these
are the same function, but we copied old versions so have two copies
of similar code) uses code that confuses gcc 12.1.1 into thinking
that a temporary function variable ("tmp" in cycle()) is saved on
a non-temporary array. The code really does this saving, but the
position saved will not be read again after the function returns.

So to avoid these false alarms we add the -Wno-dangling-pointer
option when compiling those two files.

This is one of the fixes needed to build on Fedora 36 (refs #1198).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>