virtiofs: fix incorrect use of unique_ptr
dax_window_impl::unmap() did:

   std::unique_ptr<u8> in_args {new (std::nothrow) u8[in_args_size]}

But since we use "new u8[]" here, not "new u8", we need to also delete
it with delete[], not delete, and std::unique_ptr<u8> doesn't - and
gcc 12.1.1 started detecting this and complaining.

The fix is to use std::unique_ptr<u8[]>.

This is one of the fixes needed to build on Fedora 36 (refs #1198).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>