musl: ignore a false-positive warning in __rem_pio2_large.c
Gcc 12.1.1 warns that in __rem_pio2_large() - code which we take
directly from Musl - The local variable fq[0] may be used before set.
This shouldn't happen, because the fq[] array is set by a for loop
immediately before being used. There's one snag that gcc 12.1.1
correctly caught, however - this loop will not do anything if jz,
its start, is <0. However, this is an internal function that we assume
is called with correct parameters, and "prec" with the right values
will lead to one of the numbers in init_jk[] to be used to initialze jz,
and these are all positives.

Since I don't want to modify the musl code, let's just ignore the
warning in the Makefile.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>