zfs: export symbols needed by libsolaris.so
When one builds kernel with most symbols hidden, the libsolaris.so that
depends on 92 non-libc symbols from the kernel cannot be loaded and ZFS
filesystem cannot function. To address it, this patch exposes all those
symbols by annotating then with new macro OSV_LIBSOLARIS_API which
was added to the export.h header. It also adds new file with a list of
symbols to be exported - osv_libsolaris.so.symbols.

The kernel built with most symbols hidden gets larger by 12K only.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>