zfs: make mkfs do not depend on OSv C++ api
This patch makes number of modifications across various files but mainly
it changes mkfs.cc to switch it from using osv::run() and zfs::zfsdev_init() to
new C-wrappers osv_run_app() and zfsdev_init() respectively.

In adition, it also changes both mkfs.cc and cpiod.cc to expose the symbol
main() which until now would be hidden if OSv was built with most symbols
hidden.

Finally, we modify the main makefile to build both cpiod.so and mkfs.so
as proper C++ binaries in order to make them properly express dependency on
libstdc++.so. This is necessary to make OSv dynamic linker load libstdc++.so
from the image if kernel is built to hide most symbols including the C++
standard library ones.

Refs #1186

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>