zfs: allow mounting and building on host
This requires OpenZFS installed on host (see
https://openzfs.github.io/openzfs-docs/Getting%20Started/Fedora/index.html
for Fedora and
https://openzfs.github.io/openzfs-docs/Getting%20Started/Ubuntu/index.html
for Ubuntu).

In essence this patch adds new script zfs-image-on-host.sh that allows
mounting and build ZFS images using OpenZFS without running OSv. It also
modifies the build script to support new option: '--use-openzfs' that
delegates to zfs-image-on-host.sh to build the ZFS image. Please see
examples above:

./scripts/build image=native-example fs=zfs -j$(nproc) --use-openzfs

./scripts/build image=native-example fs=rofs_with_zfs -j$(nproc) --use-openzfs

./scripts/build image=native-example fs=rofs -j$(nproc) --use-openzfs --create-zfs-disk

Fixes #1068

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>