pthread_create() - prefault top of stack
This is the original patch from Jan 11, 2021 applied manually
to resolve conflicts caused by subsequent changes to the relevant
files since.
Waldemar Kozaczuk
------------------------------------------------------------------
Since commit 695375f65303e13df1b9de798577ee9a4f8f9892, new threads start
with preemption disabled, and then immediately (in thread_main_c()) enable
preemption. However, this code runs on the thread's stack, which can be
supplied by the user. If this stack is lazily allocated (not pre-populated,
with pages to be populated only on first use), we will get a page fault
when thread_main_c() begins to run with preemption disabled.

This can happen if the application creates a thread stack with mmap()
without passing the MAP_STACK option.

Note that it is still strongly recommended to pre-populate thread
stacks by using MAP_STACK (or MAP_POPULATE), because we still haven't
committed a fix for issue #143. However, let's at least not crash
immediately and consistently, as we did before this patch.

This patch simply reads from the top byte of the user-supplied stack
before starting the thread - causing it to be faulted in (if needed)
at that time instead of when the thread starts. The first stack page
should be enough to start thread_main_c() and get to the point where
it re-enables preemption.

This patch also includes a test reproducing this issue and demonstrating
that it is fixed.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>