scripts/build-capstan-mpm-packages: enhance to build custom kernels
This patch enhances scripts/build-capstan-mpm-packages to allow building
custom kernels like in this example:

./scripts/build-capstan-mpm-packages kernel conf_hide_symbols=1 fs=rofs drivers_profile=microvm loader_image=osv-loader-microvm

In essence all extra parameters passed to script that are either
conf_*=*, fs=* or drivers_profile=* as relayed as is and passed to the
script/build downstream.

In addition one can specify the capstan image name by using the
parameter loader_image.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>