modules/httpserver-api: add missing include
An include of <map> was missing, which was somehow worked around in the
past but on Fedora 37 we really need to explicitly include it.
After this patch, "scripts/build" (building the OSv kernel and the Lua
shell) works on Fedora 37. Not all other modules work (noteably the Java
module doesn't - this is issue #1211).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>