syscall: save FPU state when setting up large syscall stack
As issue #1233 explains OSv crashes occasionally when launching an app like golang httpserver that executes a syscall instruction. The most likely cause of this is that OSv does not save and restore the FPU state as we ought to when setting up a large syscall stack which happens once per application thread.

So this patch changes both thread::setup_large_syscall_stack() and thread::free_tiny_syscall_stack() to save and restore the FPU state. In addition, it also doubles the tiny stack size to 2048 bytes to account for the extra FPU state size. Please note that in reality, OSv will allocate a whole 4K page for each tiny stack as the large malloc pool can only satisfy requests up to and equal to 1024 bytes. So this will triple the memory usage (bear in mind that only 3 allocations of 1024 bytes can fit into a single 4K page). In the future, we might think of a solution to use per-cpu tiny stacks instead but at this moment it is not clear how this would work.

Fixes #1233

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>