libc: add stub strfromf128(), strtof128()
Recent versions of libstdc++ (such as the one that comes with Fedora 38)
started using the functions strtof128() and strfromf128() when some C++
template uses the _Float128 type.

It's not trivial to implement these functions: I created in
tests/tst-f128.cc a test of what these functions need to do according
to these functions on glibc - the test passes only on glibc.
But luckily, we don't really need to implement them - it's enough
to stub them (and they'll abort if ever used by the application).

After this patch, the OSv kernel and various images (e.g., my favorite
"scripts/build image=rogue; scripts/run.py") work on Fedora 38, but
the default image "scripts/build" still doesn't work because of a
missing symbol in Lua.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>