tests: make tst-pthread-barrier.cc work with glibc
Calling pthread_barrier_init() with last 2 arguments equal to 0
is handled more leniently on OSv with OSv libc. The same call
crashes with segmentation fault when compiled with glibc.

Regardless, if that is a glibc bug, we deactivate this particular
part of the test with ifdef so that the test can be executed
on Linux and on OSv with Linux dynamic linker.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>