tests: make tst-string.cc work with glibc
The musl version of strsignal() used by OSv returns slightly different
message string than glibc. More specifically glibc appends the signal
number to the string.

We could change musl version to behave exactly like glibc, but instead
we tweak the tst-string.cc with ifdef to test correct message version
depending how it is compiled.

In the end, the test now works on Linux and on OSv with Linux dynamic
linker.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>