tst-realloc: test malloc_usable_size() only when running with OSv libc
The results returned malloc_usable_size() are very much dependant
on the malloc() implementation. Let us disable this part of the
tst-realloc.c when built with glibc.

As of this commit, most of the unit tests should pass when running on OSv
with Linux dynamic linker:

scripts/test.py --linux_ld -m modules/tests-with-linux-ld/usr.manifest \
  -d tst-futimesat \
  -d tst-kill \
  -d tst-mmap \
  -d tst-pthread-clock \
  -d tst-reloc \
  -d tst-shm \
  -d tst-sigaction \
  -d tst-sigwait \
  -d tst-stdio-rofs \
  -d tst-symlink \
  -d tst-symlink-rofs \
  -d tst-time \
  -d tst-truncate \
  -d tst-utimensat \
  -d tst-utimes \
  -d tst-wctype

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>