aarch64: set the generic timer PPI interrupt to the type level
The AArch64 Programmer's Guides Generic Timer states this in the section 3.4:

"The interrupts generated by the timer behave in a level-sensitive manner."

The type of the generic timer interrupt was set incorrectly to edge and
most likely was ignored by QEMU so OSv would work fine. But when testing
on firecracker on Rock 5B OSv would crash when initializing the timer interrupt
in arm_clock_events::arm_clock_events().

This patch fixes it by correctly setting interrupt type to level.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>