fs: add fortified openat() variants
The "fortify source" compiler feature changes calls to openat() with a
missing final mode argument to calls to __openat_2(). This function,
which OSv is missing, was supposed to check that the flag didn't ask
for O_CREAT (which requires a mode!) and then call the regular function.
This patch adds the missing function, which is needed when building
the OSv HTTP server on Fedora 39 - apparently the Boost filesystem
library we are using was built with the fortify source option.

Note that __openat_2() has nothing to do with openat2() - the name
of the former was derived from openat() similar to how __open_2()
is a 2-parameter fortified version of open(). Ironically, __openat_2()
actually has 3 parameters, not 2 :-)

After this patch, "scripts/build; scripts/run.py" still fails on
Fedora 39, because of a missing setfsuid needed from libtinfo.so.6.
But we passed the previous failures.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>