vfs: make headers export C API when included in C++
The upcoming C++ implementation of the ext4 filesystem as a module
and shared library needs some VFS headers which cannot be
included as C++ because it would pull all kind of other C++
headers (boost, etc) that we do not need and want in this ext4
implementation.

The headers included by C++ compiler are preprocessed with the __cplusplus
macro which enables them as C++. In order to make them "include-able" as
C in C++ source code, we add extra macro USE_C_INTERFACE which when
defined in C++ code would force the relevant headers to expose its C api.

This example illustrates how this can be accomplished:

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>

Closes #1309