block cache: use correct data type for block number
This issue was discovered by Jan Braunwarth when implementing
the NVMe driver and full credit goes to him.

The b_blkno field of the buf struct uses 4-bytes int type
and may cause overflow when block cache is used by bdev_read()
and bdev_write() in devfs. To correct we change the type to
off_t.

Signed-off-by: Jan Braunwarth <jan.braunwarth@gmail.com>
Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>