Remove redundant template name from method definition
When defining a method and especially a constructor in a templated class
c<T>, there is no need to use the "<T>" to "disambiguate" the name of
the method or constructors. This "disambiguation" syntax used to work but was
outlawed in C++20 (see https://stackoverflow.com/questions/63513984)
and doesn't compile with gcc 14.1.1, giving the error:

include/lockfree/queue-mpsc.hh:39:19: error: template-id not allowed for constructor in C++20 [-Werror=template-id-cdtor]
   39 |     linked_item<T>() : value(), next(nullptr) { }

This patch removes the unnecessary "disambiguations" and fixes a part of
the build on gcc 14.1.1 (e.g., on Fedora 40). There are other problems
too, which will be solved in the following patches.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>