osv::fprintf(): replace its usages with osv::sprintf()
This is another commit to remove code that directly or indirectly
depends on std::locale.

In this particular case, the usage of osv::fprintf() requires use of
string streams which we do not want to use because these implicitly pull
std::locale. So instead, we replace osv::fprintf() + std::ostringstream
with osv::sprintf().

Ref #1335

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>