add osv::split() and osv::replace_all()
The functions boost::split() and boost::replace_all() implicitly
depends on std::locale(). Therefore this patch adds the OSv
equivalent implementations of those that do not pull std::locale().

The osv::split() is heavily based on the example here -
https://cplusplus.com/faq/sequences/strings/split/#string-find_first_of

This patch also adds new unit test to verify these 2 new functions.

This is another commit to remove code that directly or indirectly
depends on std::locale.

Ref #1335

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>