build: disable JVM balloon mechanism
The JVM balloon mechanism introduced 10 years ago was designed to
work with JNI and Java 7 and 8. As the latest issue #1038 reported 5
years ago, it does not work.

Based on my limited understanding, it relies on old Java wrapper
and JNI mechanism that has since been replaced by JNA from project
Panama. On top of this the Java wrapper does not work with Java > 11
- see #1169.

The latest version of Java is 23 and I doubt this incarnation of
the JVM balloon can ever be made to work with latest Java without
significant redesigning.

In any case, we are simply disabling the relevant code in OSv kernel
by default. One can always re-enable it via new configuration knob.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>