syscalls: add build mechanism to include specific syscalls only
This patch adds new mechanism that allows including specific
syscalls based on a supplied list during the build time.

To support it, we extract most of the code in the switch statement
of the syscall() function in linux.cc into new file
syscalls/syscalls.cc.in. We also extract list of TRACEPOINT() statements
into another new file - syscalls/syscall_tracepoints.cc.in. These 2
files will be fed into a selection process handled by new script
- gen-syscalls - and included by linux.cc.

The scripts/gen-syscalls is called by main makefile in a similar
way the gen-version-header or gen-drivers-config-header is and by
default (if conf_syscalls_list_file is not set) it copies the
syscalls.cc.in and syscall_tracepoints.cc.in as is into
$(out)/gen/include/osv. If conf_syscalls_list_file is set and
points to a file with a list of syscall names, then gen-syscalls
filters in only the corresponding syscalls and their tracepoints from
syscalls.cc.in and syscall_tracepoints.cc.in into $(out)/gen/include/syscalls.cc
and $(out)/gen/include/syscall_tracepoints.cc.

For example, given a file 'native_list' with this content:

```
arch_prctl
clock_gettime
fstatat
getrandom
mprotect
prlimit64
readlink
sys_brk
sys_ioctl
sys_set_robust_list
sys_set_tid_address
write
sys_exit
sys_exit_group
```

the command:
```
./scripts/build image=native-example conf_syscalls_list_file=native_list
```

will build a kernel with 15 (14 from the list plus futex) syscalls only.

Please note that the names of syscalls must match the names in
syscalls/syscalls.cc.in. One way to get a list of those for specific
application is to run it on OSv with strace active like this:

```
./scripts/firecracker.py -e '--strace --trace=syscall* /hello-static-non-pie' | grep -o "syscall_.*(" | tr -d "(" | sort | uniq | cut -d '_' -f 2-
```

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>