support access to TLS variables located in dlopen-ed objects
The original intention of this patch was to provide a unit test
to verify handling of TLS variables in dlopen-ed shared libraries.

This does not work on aarch64 because we do not support so called
dynamic TLS descriptors yet. On x64 this functionality should work in theory
but I have found some small gaps and one important bug.

Firstly, to support dlsym() of TLS variables, we enhance the
obj::relocated_addr() to handle STT_TLS.

Secondly, we fix an important bug in the logic to set up so called
local-exec TLS. The init_static_tls() iterates over all ELF objects to
identify if any of them use TLS and setup initial- or local-exec TLS
accordingly. It turns out that if a PIE using local exec is relocated
without any DT_NEEDED dependencies using TLS, dynamic linker never marks
the object to have TLS (see `_static_tls` variable) and
init_static_tls() will skip setting local exec. This is the exact
scenario when tst-tls-pie-dlopen.so dlopen-s libtsl.so.
To fix this, we tweak the init_static_tls() to handle scenario where
initial-exec is empty but local-exec is not. We also fix
arch_relocate_rela() to call alloc_static_tls() in the cases when
the object is a dynamically linked executable - PIE.

Lastly, this patch modifies tst-tls.cc to support verifying access
of TLS variables when libtls.so is dlopen-ed and variables are resolved
using dlsym.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>