aarch64: prevent page faults when enumerating PCI devices
On some aarch64 platforms, the PCI config memory area is not large
enough to enumerate all potential 256 buses and causes the tst-hub.cc
which calls pci_devices_print() to trigger page fault. This happens
because the read_pci_config() tries to read beyond the memory-mapped
configuration area.

To fix it, this patch tweaks all read_pci_config*() and
write_pci_config*() functions to check if calculated offset is valid
and if not return error code or abort when writing.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>