fix buggy irq_spinlock
OSv prouds itself in not using spin locks except in the very few
cases when interrupts are disabled or in the early initialization
phase.

One such case is the admin queue configuration logic in the ENA driver
adapted from FreeBSD implementation. There we introduced new type of
spin lock - irq_spinlock - which disables interrupts on entry if enabled,
and enables if they were enabled on exit. This is in contrast to
the regular spin lock that disables/enables preemption.

When trying to use irq_spinlock in the initialization logic in GIC
on aarch64, I have discovered that irq_spinlock has serious logical
flaw: it saves and restores interrupt state in a variable _irq_lock
of type irq_save_lock_type which is shared across all cpus. This a problem,
because in a contended case, on entry on cpu 1 interrupts maybe be
enabled but disabled on cpu 2, in which case, on exit we may restore
wrong interrupt state on some or all cpus. The correct solution is to
save interrupt state on stack instead so that each cpu state is saved/
restored independently.

This patch replaces irq_spinlock with new np_spinlock ("np" stands for
non-preempt) that is a plain spinlock which does not disable/enable
preemption nor interrupts. Then, we also modify all uses of now defunct
irq_spinlock in the ena code to use np_spinlock along with the
irq_save_lock_type variable defined on stack.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>