libext: count references when opening and closing i-nodes
As the issue #1373 explains, current libext implementation does
not prevent file from being deleted when it is still open for
reading or writing. To fix this deficiency, this patch implements
necessary reference counting in ext_open() and ext_close() so that
when ext_dir_remove_entry() is called it will only delete i-node
if the ref_count = 0. Otherwise, it will set the flag delete_on_last_close,
so that when ext_close() executes later it will delete the-inode if the
flag is true and ref_count = 0.

There may be cases like the one caught in the tst-symlink unit test,
when the application never closes the file descriptor after removing
a file. For that reason, we store ids of the i-nodes to be deleted
when the delete_on_last_close is set and we delete those which
have never been closed when unmounting the filesystem in ext_unmount()
- see the ext_delete_outstanding_inodes() function.

This patch also cleans sone of the ext_debug() calls to make
them a little more consistent.

Fixes #1373

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>