tests: Fix bashism in java-tests Makefile
For Makefiles, we cannot assume that $SHELL is bash. Therefore, unless
we set $SHELL explicitly, we must not use any non-POSIX bash language
extensions. The `[[` operator is such a bash extension and causes the
following output on Debian-derivates where /bin/sh is dash:

	/bin/sh: 1: [: javac 1.8: unexpected operator

This commit fixes this by using POSIX's `[` operator.